Add survey result update button

+++ src/views/surveys.py
@@ -438,14 +438,18 @@ def delete_survey_result():
     """
     result_id = request.form["result_id"]
     survey_id = request.form["survey_id"]
+    action = request.form['submit']
 
-    survey_service.delete_survey_result(result_id)
+    if action == "Delete result":
+        survey_service.delete_survey_result(result_id)
 
-    # TODO: remove
-    print(f"Delete: {result_id}", flush=True)
-    print(f"Redirect to survey: {survey_id}", flush=True)
+    if action == "Update result":
+        # TODO: Implement result update
+        # survey_service.update_survey_result(result_id)
+        print(f"Update result: {result_id}", flush=True)
 
     return redirect(f"/surveys/{survey_id}/new-survey-result")
+    
         
 @surveys.route("/surveys")
 def view_surveys():
